+++ src/app/podcasts/layout.jsx
@@ -1,4 +1,6 @@
 import localFont from 'next/font/local'
+import Header from '@/components/general/Header'
+import Footer from '@/components/general/Footer'
 
 const satoshi = localFont({
   src: [{ path: '../../assets/podcasts/fonts/satoshi.woff2', weight: '300 900', style: 'normal' }],
@@ -10,7 +12,9 @@ const satoshi = localFont({
 export default function LayoutWrapper({ children }) {
   return (
     <section data-app="podcasts" className={satoshi.variable + ' flex min-h-full'}>
-      <div className="w-full">{children}</div>
+      <div className="w-full">
+        {children}
+      </div>
     </section>
   )
 }
